backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
redis==5.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.32.3
openai==1.42.0

